added commonaction
commonaction is the php cornerstone of the backend. it executes every time a php file is loaded and sets some session file. it also manages pages visibility it must be incuded at the top of every action.php file as it is and abstract methode. the function executeAction must be defined in all files. this methode executes after visibility and session check