making sign in work
created SignInAction that checks to see if username and password entered is correct. if correct, a variable is updated. the line is taken from the database table and if it isn't empty, make visibility from public to member.  the SignIn file made during phase 1 was also updated with a function to display an error message incase wrong password or username is entered. An authenticate function was also written in MySQLrequests that returns username and password from the database. technically, this would count for 3 commits but i did all the changes as one commit.