query posts from database onto tabmenu
Depending on what tab is clicked on, the server requests a query through the database for the relevant posts (most liked, newest, and least liked).

some problems were run into (had difficulty querying the post creator - which could be done through querying the  PostID for the post_creator).

Will consult on how to query the number of replies given to post in order to finish the unanswered tab (it is filled by querying the leasted liked posts at the moment).