Profile page made, and several refactoring
refactoring of the post question page to make for a cleaner look and more fluild submission to the database.

Made profilePageAction.php which allows the profile page to work
Made editProfilePage so that use is allowed to change their username and switch out profile pictures

Partial header is updated to include profile page services like change picture, change password and redirect to main profile page.