sign in user through facebook
create a facebook app as developer and include the sdk pachage to the repo.
update sign in code with new facebook sign in code.
update sign in action to include code that first checks to see if the users facebook sign in info. is present in our database. if not present, add it newly. else sign in user into the website.
refactoring sign up code.